Contrail to support 4 byte AS
Contrail only supports 2 byte AS today. There is a requirement to extend
it to use 4 byte AS also.
This is first draft to support the same. Some testcaes are added and
some more may need to be added.

Partial-Bug: #1416551

+++ src/bgp/bgp_config.h
@@ -339,7 +339,7 @@ struct StaticRouteConfig {
     std::vector<std::string> communities;
 };
 
-typedef std::vector<uint16_t> AsnList;
+typedef std::vector<uint32_t> AsnList;
 typedef std::vector<std::string> CommunityList;
 typedef std::vector<std::string> ProtocolList;
